+++ Locales/Locale.lua
@@ -1,0 +1,5 @@
+local ADDON, T = ...
+
+T.L = setmetatable({}, {
+    __index = function(_, key) return key end,
+})